feat: move contact info tab content into profile tab
