+++ src/components/config/channels/whatsapp/Config.vue
@@ -43,11 +43,6 @@
         @save="() => fetchData({ skipLoad: true })"
       />
 
-      <template slot="tab-head-contact_info">
-        {{ $t('WhatsApp.config.tabs.contact_info') }}
-      </template>
-      <ContactInfoTab slot="tab-panel-contact_info" :app="app" @close="closeConfig" />
-
       <template slot="tab-head-webhook_info">
         {{ $t('WhatsApp.config.tabs.webhook_info') }}
       </template>
@@ -65,7 +60,6 @@
 <script>
   import AccountTab from './components/tabs/AccountTab.vue';
   import ProfileTab from './components/tabs/ProfileTab.vue';
-  import ContactInfoTab from './components/tabs/ContactInfoTab.vue';
   import ConversationsTab from './components/tabs/ConversationsTab.vue';
   import WebhookTab from './components/tabs/WebhookTab.vue';
   import skeletonLoading from './loadings/Config.vue';
@@ -79,7 +73,6 @@
       skeletonLoading,
       AccountTab,
       ProfileTab,
-      ContactInfoTab,
       ConversationsTab,
       WebhookTab,
     },
@@ -118,7 +111,7 @@
         errorCurrentApp: (state) => state.appType.errorCurrentApp,
       }),
       configTabs() {
-        return ['account', 'profile', 'contact_info', 'webhook_info', 'conversations'];
+        return ['account', 'profile', 'webhook_info', 'conversations'];
       },
       documentationLink() {
         return this.documentations[this.$i18n.locale] ?? this.documentations['en-us'];
