+++ src/components/config/channels/whatsapp/components/tabs/ProfileTab.vue
@@ -1,6 +1,19 @@
 <template>
   <div class="profile-content">
-    <dynamic-form class="profile-content__form" :inputs="profileInputs" @input="updateInputs" />
+    <div v-if="!loadingContactInfo" class="profile-content__wrapper">
+      <dynamic-form
+        class="profile-content__form"
+        :inputs="profileInputs"
+        @input="updateProfileInputs"
+      />
+
+      <dynamic-form
+        class="contact-info__form"
+        :inputs="contactInfoInputs"
+        @input="updateContactInfoInputs"
+      />
+    </div>
+    <skeleton-loading v-else />
 
     <div class="profile-content__buttons">
       <unnnic-button
@@ -15,8 +28,9 @@
         class="profile-content__buttons__save"
         type="secondary"
         size="large"
+        :disabled="!modifiedInputs"
         :text="$t('WhatsApp.config.profile.save_changes')"
-        @click="saveProfile"
+        @click="handleSave"
       />
     </div>
   </div>
@@ -24,14 +38,15 @@
 
 <script>
   import DynamicForm from '@/components/config/DynamicForm';
+  import skeletonLoading from './loadings/ProfileTab.vue';
   import removeEmpty from '@/utils/clean.js';
   import { toBase64, getHeightAndWidthFromDataUrl } from '@/utils/files.js';
-  import { mapActions, mapState } from 'vuex';
+  import { mapActions, mapState, mapGetters } from 'vuex';
   import { unnnicCallAlert } from '@weni/unnnic-system';
 
   export default {
     name: 'ProfileTab',
-    components: { DynamicForm },
+    components: { DynamicForm, skeletonLoading },
     props: {
       app: {
         type: Object,
@@ -42,17 +57,40 @@
         default: /* istanbul ignore next */ () => {},
       },
     },
-    mounted() {
+    async mounted() {
       /* istanbul ignore next */
       if (this.app.code !== 'wpp') {
         this.profileInputs = this.profileInputs.filter((value) => {
           return value.name !== 'status';
         });
       }
+
+      if (!this.fetchedContactInfo) {
+        try {
+          await this.fetchWppContactInfo({ code: this.app.code, appUuid: this.app.uuid });
+        } catch (err) {
+          unnnicCallAlert({
+            props: {
+              text: this.$t('WhatsApp.config.error.data_fetch'),
+              title: this.$t('general.error'),
+              icon: 'alert-circle-1-1',
+              scheme: 'feedback-red',
+              position: 'bottom-right',
+              closeText: this.$t('general.Close'),
+            },
+            seconds: 8,
+          });
+        }
+      }
+      this.setInitialInputs();
+
+      setTimeout(() => (this.hadInitialModified = true), 500);
     },
     data() {
       return {
         modifiedInitialPhoto: false,
+        modifiedInputs: false,
+        hadInitialModified: false,
         profileInputs: [
           {
             type: 'upload',
@@ -100,11 +138,50 @@
             }),
           },
         ],
+        contactInfoInputs: [
+          {
+            type: 'input',
+            name: 'websites.0',
+            label: 'WhatsApp.config.contact_info.website_url.label',
+            value: null,
+          },
+          {
+            name: 'websites.1',
+            label: null,
+            value: null,
+            type: 'input',
+          },
+          {
+            type: 'input',
+            name: 'email',
+            label: 'WhatsApp.config.contact_info.email.label',
+            placeholder: 'WhatsApp.config.contact_info.email.placeholder',
+            value: null,
+            validator: (value) => {
+              if (!value) return true;
+
+              const emailRegex = /^\w+([-+.']\w+)*@\w+([-.]\w+)*\.\w+([-.]\w+)*$/;
+              return emailRegex.test(value);
+            },
+            error: false,
+            errorMessage: 'WhatsApp.config.error.invalid_email',
+          },
+          {
+            type: 'input',
+            name: 'address',
+            label: 'WhatsApp.config.contact_info.address.label',
+            placeholder: 'WhatsApp.config.contact_info.address.placeholder',
+            value: null,
+          },
+        ],
       };
     },
     watch: {
       profileInputs: {
         handler: function (val) {
+          if (this.hadInitialModified) {
+            this.modifiedInputs = true;
+          }
           const photoIndex = val.indexOf(val.find((input) => input.name === 'profile_image'));
           if (val[photoIndex].value[0]?.lastModified !== this.profile.photoFile?.lastModified) {
             this.modifiedInitialPhoto = true;
@@ -112,14 +189,57 @@
         },
         deep: true,
       },
+      contactInfoInputs: {
+        handler: function () {
+          if (this.hadInitialModified) {
+            this.modifiedInputs = true;
+          }
+        },
+        deep: true,
+      },
     },
     computed: {
       ...mapState('WhatsApp', ['errorUpdateWhatsAppProfile', 'errorDeleteWhatsAppProfilePhoto']),
+      ...mapGetters({
+        loadingContactInfo: 'WhatsApp/loadingContactInfo',
+        fetchedContactInfo: 'WhatsApp/fetchedContactInfo',
+        contactInfo: 'WhatsApp/contactInfo',
+      }),
     },
     methods: {
-      ...mapActions('WhatsApp', ['updateWppProfile', 'deleteWppProfilePhoto']),
-      updateInputs(inputData) {
+      ...mapActions('WhatsApp', [
+        'updateWppProfile',
+        'deleteWppProfilePhoto',
+        'fetchWppContactInfo',
+        'updateWppContactInfo',
+      ]),
+      updateProfileInputs(inputData) {
         this.profileInputs[inputData.index].value = inputData.value;
+      },
+      updateContactInfoInputs(inputData) {
+        const input = this.contactInfoInputs[inputData.index];
+        if (input.validator) {
+          this.contactInfoInputs[inputData.index].error = !input.validator(inputData.value);
+        }
+        this.contactInfoInputs[inputData.index].value = inputData.value;
+      },
+      setInitialInputs() {
+        if (this.contactInfo) {
+          Object.entries(this.contactInfo).forEach(([key, value]) => {
+            if (Array.isArray(value)) {
+              this.contactInfoInputs.forEach((input, index) => {
+                if (input.name.includes(key)) {
+                  input.value = value[index];
+                }
+              });
+            } else {
+              const input = this.contactInfoInputs.find((input) => input.name === key);
+              if (input) {
+                input.value = value;
+              }
+            }
+          });
+        }
       },
       async isValidPhotoSize(b64ProfilePhoto) {
         const { height, width } = await getHeightAndWidthFromDataUrl(b64ProfilePhoto);
@@ -140,12 +260,19 @@
         }
         return true;
       },
-      getInputValue(inputName) {
+      getProfileInputValue(inputName) {
         return this.profileInputs.find((input) => input.name === inputName)?.value;
+      },
+      getContactInfoInputValue(inputName) {
+        return this.contactInfoInputs.find((input) => input.name === inputName).value;
+      },
+      handleSave() {
+        this.saveProfile();
+        this.saveContactInfo();
       },
       async saveProfile() {
         try {
-          const photo = this.getInputValue('profile_image')[0];
+          const photo = this.getProfileInputValue('profile_image')[0];
 
           if (!photo) {
             const data = { code: this.app.code, appUuid: this.app.uuid };
@@ -165,10 +292,10 @@
 
           const payload = {
             photo: this.modifiedInitialPhoto ? b64ProfilePhoto : null,
-            status: this.getInputValue('status'),
+            status: this.getProfileInputValue('status'),
             business: {
-              description: this.getInputValue('description'),
-              vertical: this.getInputValue('sector'),
+              description: this.getProfileInputValue('description'),
+              vertical: this.getProfileInputValue('sector'),
             },
           };
 
@@ -184,7 +311,7 @@
           unnnicCallAlert({
             props: {
               text: this.$t('WhatsApp.config.success.profile_updated'),
-              title: 'Success',
+              title: this.$t('general.success'),
               icon: 'check-circle-1-1',
               scheme: 'feedback-green',
               position: 'bottom-right',
@@ -196,7 +323,61 @@
           unnnicCallAlert({
             props: {
               text: this.$t('apps.details.status_error'),
-              title: 'Error',
+              title: this.$t('general.error'),
+              icon: 'alert-circle-1-1',
+              scheme: 'feedback-red',
+              position: 'bottom-right',
+              closeText: this.$t('general.Close'),
+            },
+            seconds: 6,
+          });
+        }
+      },
+      async saveContactInfo() {
+        if (this.contactInfoInputs.some((input) => input.error)) {
+          unnnicCallAlert({
+            props: {
+              text: this.$t('WhatsApp.config.error.invalid_data'),
+              title: this.$t('general.error'),
+              icon: 'alert-circle-1-1',
+              scheme: 'feedback-red',
+              position: 'bottom-right',
+              closeText: this.$t('general.Close'),
+            },
+            seconds: 3,
+          });
+          return;
+        }
+
+        try {
+          const payload = {
+            websites: Array.of(
+              this.getContactInfoInputValue('websites.0'),
+              this.getContactInfoInputValue('websites.1'),
+            ),
+            email: this.getContactInfoInputValue('email'),
+            address: this.getContactInfoInputValue('address'),
+          };
+
+          const data = removeEmpty({ code: this.app.code, appUuid: this.app.uuid, payload });
+          await this.updateWppContactInfo(data);
+
+          unnnicCallAlert({
+            props: {
+              text: this.$t('WhatsApp.config.success.contact_info_update'),
+              title: this.$t('general.success'),
+              icon: 'check-circle-1-1',
+              scheme: 'feedback-green',
+              position: 'bottom-right',
+              closeText: this.$t('general.Close'),
+            },
+            seconds: 3,
+          });
+        } catch (err) {
+          unnnicCallAlert({
+            props: {
+              text: this.$t('WhatsApp.config.error.data_update'),
+              title: this.$t('general.error'),
               icon: 'alert-circle-1-1',
               scheme: 'feedback-red',
               position: 'bottom-right',
@@ -217,12 +398,18 @@
     gap: $unnnic-inline-xs;
     height: 100%;
     margin-top: $unnnic-spacing-stack-xs;
+    overflow: hidden;
+
+    &__wrapper {
+      display: flex;
+      flex-direction: column;
+      height: 100%;
+      overflow: auto;
+    }
 
     &__form {
       display: flex;
       flex-direction: column;
-      overflow-x: hidden;
-      flex: 1;
       gap: $unnnic-spacing-inline-xs;
 
       padding-right: $unnnic-spacing-inline-xs;
@@ -244,6 +431,26 @@
       font-family: $unnnic-font-family-secondary;
       font-size: $unnnic-font-size-body-lg;
       line-height: $unnnic-font-size-body-lg + $unnnic-line-height-md;
+    }
+  }
+
+  .contact-info {
+    &__form {
+      display: flex;
+      flex-direction: column;
+      gap: $unnnic-spacing-inline-xs;
+
+      padding-right: $unnnic-spacing-inline-xs;
+    }
+
+    &__buttons {
+      margin-top: $unnnic-spacing-stack-sm;
+      display: flex;
+
+      &__cancel,
+      &__save {
+        flex-grow: 1;
+      }
     }
   }
 </style>
